got person follower working
