+++ scripts/person_follower.py
@@ -1,141 +1,59 @@
 #!/usr/bin/env python2
 
-from __future__ import division
-
-from sensor_msgs.msg import PointCloud
-from geometry_msgs.msg import Point, Vector3, Pose
-from std_msgs.msg import Header, ColorRGBA
-from visualization_msgs.msg import Marker
 import rospy
-import tf
-import time
+from visualization_msgs.msg import Marker
+from geometry_msgs.msg import Pose, Point, Twist, Vector3
+import math
+
 
 class PersonFollower(object):
     def __init__(self):
         super(PersonFollower, self).__init__()
-        rospy.init_node('person_follower')
-        self.listener = tf.TransformListener(10)
+        rospy.init_node('perosn_follower')
+        rospy.Subscriber('/visualization_messages/Centroid', Marker, self.get_target)
+        self.publisher = rospy.Publisher('/cmd_vel', Twist, queue_size=10)
 
-        rospy.Subscriber('/projected_stable_scan', PointCloud, self.process_scan)
-        self.box_publisher = rospy.Publisher('/visualization_messages/Box', Marker, queue_size=10)
-        self.point_publisher = rospy.Publisher('/visualization_messages/Marker', Marker, queue_size=10)
-        self.centroid_publisher = rospy.Publisher('/visualization_messages/Centroid', Marker, queue_size=10)
-        
+        self.target = None
         self.got_first_message = False
-        self.scan_msg = None
-        self.detection_length = 1
-        self.detection_width = 0.5
 
-    def process_scan(self, msg):
-        self.scan_msg = msg
+        rospy.on_shutdown(self.stop)
+
+    def get_target(self, msg):
+        self.target = msg.pose.position
 
         if not self.got_first_message:
             self.got_first_message = True
 
-    def publish_box(self, length, width):
-        far_left = Point(length, -1 * width/2.0, 0)
-        far_right = Point(length,     width/2.0, 0)
-        close_left = Point(0, -1 * width/2.0, 0)
-        close_right = Point(0, width/2.0, 0)
-        box = [far_left, far_right, close_right, close_left, far_left]
-        marker = Marker(
-            type=Marker.LINE_STRIP,
-            header=Header(
-                stamp=rospy.Time.now(),
-                frame_id='base_link'
-            ),
-            points=box,
-            scale=Vector3(0.01, 0.2, 0.2),
-            color=ColorRGBA(0.0, 1.0, 0.0, 1.0)
-        )
-        self.box_publisher.publish(marker)
-
-    def publish_points(self, points):
-        marker = Marker(
-            type=Marker.POINTS,
-            header=Header(
-                stamp=rospy.Time.now(),
-                frame_id='base_link'
-            ),
-            points=points,
-            scale=Vector3(0.05, 0.05, 0.05),
-            color=ColorRGBA(1.0, 1.0, 0.0, 1.0)
-        )
-        self.point_publisher.publish(marker)
-
-    def publish_centroid(self, point):
-        marker = Marker(
-            type=Marker.SPHERE,
-            header=Header(
-                stamp=rospy.Time.now(),
-                frame_id='base_link'
-            ),
-            pose=Pose(
-                position=point
-            ),
-            scale=Vector3(0.05, 0.05, 0.05),
-            color=ColorRGBA(0.0, 1.0, 1.0, 1.0)
-        )
-        self.centroid_publisher.publish(marker)
-
-    def in_box(self, point, box_length, box_width):
-
-        # The x direction in baselink points forward from the robot
-        x_lower_lim = 0
-        x_upper_lim = box_length
-        y_lower_lim = -1 * box_width/2.0
-        y_upper_lim = box_width/2.0
-
-        in_x_bounds = point.x > x_lower_lim and point.x < x_upper_lim
-        in_y_bounds = point.y > y_lower_lim and point.y < y_upper_lim
-        return in_x_bounds and in_y_bounds
-
-    def to_base_link(self, msg):
-        now = rospy.Time.now()
-
-        new_msg = PointCloud(
-            header=Header(
-                stamp=now, 
-                frame_id='odom'
-            ), 
-            points = msg.points)
-        try:
-            self.listener.waitForTransform('base_link', 'odom', now, rospy.Duration(4.0))
-            return self.listener.transformPointCloud('base_link', new_msg)
-        except (tf.LookupException, tf.ConnectivityException):
-            print 'Time lookup Exception!'
-
-    def get_centroid(self, points):
-        if len(points) == 0:
-            return None
-
-        return Point(
-            x = sum([p.x for p in points]) / len(points),
-            y = sum([p.y for p in points]) / len(points),
-            z = sum([p.z for p in points]) / len(points)
+    def stop(self):
+        self.publisher.publish(
+            Twist(linear=Vector3(0.0, 0.0, 0.0), angular=Vector3(0.0, 0.0, 0.0))
         )
 
     def run(self):
-        r = rospy.Rate(50)
+        r = rospy.Rate(40)
 
         while not self.got_first_message:
             r.sleep()
 
-        time.sleep(1)
-
         while not rospy.is_shutdown():
-            in_scan_range = lambda p: self.in_box(p, self.detection_length, self.detection_width)
-            scan_msg_base_link = self.to_base_link(self.scan_msg)
-            
-            points_in_range = filter(in_scan_range, scan_msg_base_link.points)
+            angle = math.atan2(self.target.y, self.target.x)
+            distance = math.sqrt(self.target.y**2 + self.target.x**2)
+            print 'distance: ' + str(distance)
+            print 'angle: ' + str(angle)
 
-            self.publish_points(points_in_range)
-            self.publish_box(self.detection_length, self.detection_width)
+            optimal_dist = 0.5
+            # This makes it so robot doesn't back/forward if there is no
+            # object detected.
+            if distance == 0.0:
+                forward_vec = Vector3(0.0, 0.0, 0.0)
+            else:
+                forward_vec = Vector3(distance - optimal_dist, 0.0, 0.0)
 
-            self.publish_centroid(
-                self.get_centroid(points_in_range)
+            turn_msg = Twist(
+                linear=forward_vec,
+                angular=Vector3(0.0, 0.0, angle)
             )
-
-            r.sleep()
+            self.publisher.publish(turn_msg)
 
 PersonFollower().run()
+
